@@ -1,7 +1,7 @@
 package com.djay.locatesecurly.ui;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.MenuItem;
 
 import com.djay.locatesecurly.db.DatabaseManager;
