@@ -1,7 +1,7 @@
 package com.djay.locatesecurly.utils;
 
-import android.support.annotation.IntDef;
-import android.support.annotation.StringRes;
+import androidx.annotation.IntDef;
+import androidx.annotation.StringRes;
 import android.widget.Toast;
 
 import com.djay.locatesecurly.App;
